pyproject.toml
[tool.poetry]
name = "nonebot-plugin-templates-draw"
version = "1.0.0"
description = "NoneBot2 一个模板绘图插件"
authors = ["padoru233"]
readme = "README.md"
packages = [{include = "nonebot_plugin_templates_draw"}]
homepage = "https://github.com/padoru233/nonebot-plugin-templates-draw"
repository = "https://github.com/padoru233/nonebot-plugin-templates-draw"
keywords = ["nonebot2", "nonebot-plugin-templates-draw"]

[tool.poetry.dependencies]
python = ">=3.9"
nonebot2 = ">=2.2.0"
nonebot-adapter-onebot = "^2.0.0"
nonebot-plugin-localstore = "^0.7.1"

httpx = ">=0.27.2, <1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
select = [
    "F",
    "E",
    "W",
    "I",
    "UP",
    "RUF"
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "UP006",  # non-pep585-annotation
    "UP035",  # deprecated-import
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
]
